move kick_adns_connection from dnskey to initiate
