+++ programs/pluto/dnskey.c
@@ -47,6 +47,7 @@
 #include "keys.h"	    /* needs connections.h */
 #include "dnskey.h"
 #include "packet.h"
+#include "state.h"
 #include "timer.h"
 #include "pluto/server.h"
 
@@ -1845,23 +1846,6 @@ void dump_addr_info(struct addrinfo *ans)
     }
 }
 
-bool kick_adns_connection(struct connection *c)
-{
-    bool kicknow = FALSE;
-
-    /* arrange to rekey the phase 1, if there was one. */
-    if (c->spd.that.host_type == KH_IPHOSTNAME) {
-        kicknow = kicknow || kick_adns_connection_lookup(c, &c->spd.that);
-    }
-    if (!kicknow && c->spd.this.host_type == KH_IPHOSTNAME) {
-        kicknow = kicknow || kick_adns_connection_lookup(c, &c->spd.this);
-    }
-    if(kicknow) {
-        restart_connections_by_peer(c);
-    }
-
-    return kicknow;
-}
 
 bool kick_adns_connection_lookup(struct connection *c
                                  , struct end *end)
